pyproject.toml
[tool.poetry]
name = "spend-time-together"
version = "0.1.0"
description = ""
authors = ["Maksim Murzin <wow_murzin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
fastapi = "^0.115.0"
uvicorn = "^0.35.0"
python-dotenv = "^1.1.0"
orjson = "^3.9.15"
sqlalchemy = "^2.0.25"
asyncmy = "^0.2.9"
pydantic-settings = "^2.1.0"
dependency-injector = "^4.41.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
factory-boy = "^3.3.0"
aiosqlite = "^0.19.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"